requirements.txt
unidecode
dedupe-variable-name
